Use logging helper for all solvers implementations
