+++ crates/solvers/src/infra/config/dex/file.rs
@@ -1,7 +1,10 @@
 //! Configuration parameters that get shared across all dex solvers.
 
 use {
-    crate::domain::{dex::slippage, eth},
+    crate::{
+        domain::{dex::slippage, eth},
+        infra::config::unwrap_or_log,
+    },
     bigdecimal::BigDecimal,
     serde::{de::DeserializeOwned, Deserialize},
     serde_with::serde_as,
@@ -64,22 +67,4 @@ pub async fn load<T: DeserializeOwned>(path: &Path) -> (super::Config, T) {
     };
 
     (config, dex)
-}
-
-/// Unwraps result or logs a `TOML` parsing error.
-fn unwrap_or_log<T, E, P>(result: Result<T, E>, path: &P) -> T
-where
-    E: Debug,
-    P: Debug,
-{
-    result.unwrap_or_else(|err| {
-        if std::env::var("TOML_TRACE_ERROR").is_ok_and(|v| v == "1") {
-            panic!("failed to parse TOML config at {path:?}: {err:#?}")
-        } else {
-            panic!(
-                "failed to parse TOML config at: {path:?}. Set TOML_TRACE_ERROR=1 to print \
-                 parsing error but this may leak secrets."
-            )
-        }
-    })
 }
