+++ crates/solvers/src/infra/config/legacy.rs
@@ -1,7 +1,7 @@
 use {
     crate::{
         domain::{eth, solver::legacy},
-        infra::contracts,
+        infra::{config::unwrap_or_log, contracts},
         util::serialize,
     },
     reqwest::Url,
@@ -36,8 +36,7 @@ pub async fn load(path: &Path) -> legacy::Config {
     let data = fs::read_to_string(path)
         .await
         .unwrap_or_else(|e| panic!("I/O error while reading {path:?}: {e:?}"));
-    let config = toml::de::from_str::<Config>(&data)
-        .unwrap_or_else(|err| panic!("TOML syntax error while reading {path:?} {err:?}"));
+    let config = unwrap_or_log(toml::de::from_str::<Config>(&data), &path);
     let contracts = contracts::Contracts::for_chain(config.chain_id);
 
     legacy::Config {
